create a flash module and port over application logic from service + change namespace + create additional tests
